Expose typed remote transfer selection context
