+++ crates/libs/lxmf-sdk/src/backend/zmq_pipeline/tests/propagation.rs
@@ -147,6 +147,8 @@ fn propagation_remote_lifecycle_uses_zmq_sdk_envelopes_and_preserves_raw_state()
                         "remote": "remote-a",
                         "propagation": {
                             "state_name": "completed",
+                            "selected_node": "node-fetch",
+                            "selected_peer": "peer-fetch",
                             "sync_progress": 1.0,
                             "last_sync_started": 1_700_001_000,
                             "last_sync_completed": 1_700_001_120,
@@ -175,6 +177,8 @@ fn propagation_remote_lifecycle_uses_zmq_sdk_envelopes_and_preserves_raw_state()
                         "remote": "remote-a",
                         "propagation": {
                             "state_name": "failed",
+                            "selected_node": "node-download",
+                            "selected_peer": "peer-download",
                             "last_sync_error": "remote download postponed",
                             "last_sync_started": 1_700_001_800,
                             "last_sync_completed": null,
@@ -207,6 +211,8 @@ fn propagation_remote_lifecycle_uses_zmq_sdk_envelopes_and_preserves_raw_state()
                         "peer": "peer-a",
                         "propagation": {
                             "state_name": "failed",
+                            "selected_node": "node-sync",
+                            "selected_peer": "peer-sync",
                             "last_sync_error": "remote sync timed out",
                             "last_sync_started": 1_700_002_000,
                             "last_sync_completed": null,
@@ -263,6 +269,8 @@ fn propagation_remote_lifecycle_uses_zmq_sdk_envelopes_and_preserves_raw_state()
                         },
                         "propagation": {
                             "state_name": "failed",
+                            "selected_node": "node-unpeer",
+                            "selected_peer": "peer-unpeer",
                             "last_sync_error": "remote unpeer denied",
                             "last_sync_started": 1_700_002_600,
                             "last_sync_completed": null,
@@ -352,6 +360,8 @@ fn propagation_remote_lifecycle_uses_zmq_sdk_envelopes_and_preserves_raw_state()
     assert_eq!(fetch.transfer_state.imported_ids, vec!["id-a".to_string(), "id-b".to_string()]);
     assert_eq!(fetch.transfer_state.transferred_bytes, 128);
     assert_eq!(fetch.transfer_state.state_name.as_deref(), Some("completed"));
+    assert_eq!(fetch.transfer_state.selected_node.as_deref(), Some("node-fetch"));
+    assert_eq!(fetch.transfer_state.selected_peer.as_deref(), Some("peer-fetch"));
     assert_eq!(fetch.transfer_state.sync_progress, Some(1.0));
     assert_eq!(fetch.transfer_state.last_sync_started, Some(1_700_001_000));
     assert_eq!(fetch.transfer_state.last_sync_completed, Some(1_700_001_120));
@@ -365,6 +375,8 @@ fn propagation_remote_lifecycle_uses_zmq_sdk_envelopes_and_preserves_raw_state()
     assert!(download.transfer_state.postponed);
     assert_eq!(download.transfer_state.postpone_reason.as_deref(), Some("timeout"));
     assert_eq!(download.transfer_state.state_name.as_deref(), Some("failed"));
+    assert_eq!(download.transfer_state.selected_node.as_deref(), Some("node-download"));
+    assert_eq!(download.transfer_state.selected_peer.as_deref(), Some("peer-download"));
     assert_eq!(download.transfer_state.failure_kind.as_deref(), Some("timeout"));
     assert!(download.transfer_state.timed_out);
     assert!(!download.transfer_state.access_denied);
@@ -386,6 +398,8 @@ fn propagation_remote_lifecycle_uses_zmq_sdk_envelopes_and_preserves_raw_state()
     assert!(sync.transfer_state.postponed);
     assert_eq!(sync.transfer_state.postpone_reason.as_deref(), Some("timeout"));
     assert_eq!(sync.transfer_state.state_name.as_deref(), Some("failed"));
+    assert_eq!(sync.transfer_state.selected_node.as_deref(), Some("node-sync"));
+    assert_eq!(sync.transfer_state.selected_peer.as_deref(), Some("peer-sync"));
     assert_eq!(sync.transfer_state.failure_kind.as_deref(), Some("timeout"));
     assert!(sync.transfer_state.timed_out);
     assert!(!sync.transfer_state.access_denied);
@@ -407,6 +421,8 @@ fn propagation_remote_lifecycle_uses_zmq_sdk_envelopes_and_preserves_raw_state()
     assert!(!unpeer.transfer_state.synced);
     assert!(!unpeer.transfer_state.postponed);
     assert_eq!(unpeer.transfer_state.state_name.as_deref(), Some("failed"));
+    assert_eq!(unpeer.transfer_state.selected_node.as_deref(), Some("node-unpeer"));
+    assert_eq!(unpeer.transfer_state.selected_peer.as_deref(), Some("peer-unpeer"));
     assert_eq!(unpeer.transfer_state.failure_kind.as_deref(), Some("no_access"));
     assert!(!unpeer.transfer_state.timed_out);
     assert!(unpeer.transfer_state.access_denied);
@@ -1068,128 +1084,6 @@ fn propagation_recovery_state_projects_status_for_zmq_sdk_clients() {
             "payload": {},
             "extensions": {}
         }))
-    );
-    server.join().expect("server joined");
-}
-
-#[test]
-fn propagation_local_payload_ingest_and_fetch_use_zmq_sdk_envelopes() {
-    let command_endpoint = unused_loopback_endpoint();
-    let response_endpoint = unused_loopback_endpoint();
-    let captured = Arc::new(Mutex::new(Vec::new()));
-    let server = spawn_response_sequence_zmq_server(
-        command_endpoint.clone(),
-        vec![
-            json!({
-                "response": {
-                    "operation_id": "app.propagation.ingest",
-                    "kind": "result",
-                    "accepted": true,
-                    "correlation_id": null,
-                    "payload": {
-                        "ingested_count": 1,
-                        "duplicate_count": 0,
-                        "payload_bytes": 18,
-                        "transferred_bytes": 18,
-                        "transient_id": "transient-sdk-ingest",
-                        "propagation": {
-                            "enabled": true,
-                            "selected_node": "router-ingest",
-                            "sync_state": 1,
-                            "state_name": "queued",
-                            "queue_depth": 5,
-                            "total_ingested": 11,
-                            "last_ingest_count": 1
-                        }
-                    }
-                }
-            }),
-            json!({
-                "response": {
-                    "operation_id": "app.propagation.fetch",
-                    "kind": "result",
-                    "accepted": true,
-                    "correlation_id": null,
-                    "payload": {
-                        "transient_id": "transient-sdk-ingest",
-                        "payload_hex": "70726f7061676174696f6e2d7061796c6f6164",
-                        "payload_bytes": 18,
-                        "transferred_bytes": 18,
-                        "propagation": {
-                            "enabled": true,
-                            "selected_node": "router-fetch",
-                            "sync_state": 2,
-                            "state_name": "serving",
-                            "queue_depth": 4,
-                            "client_propagation_messages_served": 1
-                        }
-                    }
-                }
-            }),
-        ],
-        Arc::clone(&captured),
-    );
-    let mut config = ZmqPipelineBackendConfig::local_tcp(command_endpoint, response_endpoint);
-    config.request_timeout = std::time::Duration::from_secs(2);
-    let client = ZmqPipelineBackendClient::new(config).expect("zmq client");
-
-    let ingested = client
-        .propagation_ingest(crate::PropagationIngestRequest {
-            transient_id: Some("transient-sdk-ingest".to_string()),
-            payload_hex: Some("70726f7061676174696f6e2d7061796c6f6164".to_string()),
-        })
-        .expect("propagation ingest");
-    let fetched = client
-        .propagation_fetch(crate::PropagationFetchRequest {
-            transient_id: "transient-sdk-ingest".to_string(),
-        })
-        .expect("propagation fetch");
-
-    assert_eq!(ingested.ingested_count, 1);
-    assert_eq!(ingested.duplicate_count, 0);
-    assert_eq!(ingested.payload_bytes, 18);
-    assert_eq!(ingested.transient_id, "transient-sdk-ingest");
-    assert_eq!(ingested.propagation["selected_node"], json!("router-ingest"));
-    assert_eq!(ingested.recovery_state.selected_node.as_deref(), Some("router-ingest"));
-    assert_eq!(ingested.recovery_state.sync_state, 1);
-    assert_eq!(ingested.recovery_state.state_name.as_deref(), Some("queued"));
-    assert_eq!(ingested.recovery_state.queue_depth, 5);
-    assert_eq!(ingested.recovery_state.total_ingested, 11);
-    assert_eq!(ingested.recovery_state.last_ingest_count, 1);
-    assert_eq!(fetched.transient_id, "transient-sdk-ingest");
-    assert_eq!(fetched.payload_hex, "70726f7061676174696f6e2d7061796c6f6164");
-    assert_eq!(fetched.transferred_bytes, 18);
-    assert_eq!(fetched.propagation["selected_node"], json!("router-fetch"));
-    assert_eq!(fetched.recovery_state.selected_node.as_deref(), Some("router-fetch"));
-    assert_eq!(fetched.recovery_state.sync_state, 2);
-    assert_eq!(fetched.recovery_state.state_name.as_deref(), Some("serving"));
-    assert_eq!(fetched.recovery_state.queue_depth, 4);
-    assert_eq!(fetched.recovery_state.client_propagation_messages_served, 1);
-
-    let captured = captured.lock().expect("captured requests");
-    let operation_ids = captured
-        .iter()
-        .map(|request| request.params.as_ref().expect("params")["operation_id"].clone())
-        .collect::<Vec<_>>();
-    assert_eq!(
-        operation_ids,
-        vec![json!("app.propagation.ingest"), json!("app.propagation.fetch")]
-    );
-    let kinds = captured
-        .iter()
-        .map(|request| request.params.as_ref().expect("params")["kind"].clone())
-        .collect::<Vec<_>>();
-    assert_eq!(kinds, vec![json!("command"), json!("command")]);
-    assert_eq!(
-        captured[0].params.as_ref().expect("params")["payload"],
-        json!({
-            "transient_id": "transient-sdk-ingest",
-            "payload_hex": "70726f7061676174696f6e2d7061796c6f6164"
-        })
-    );
-    assert_eq!(
-        captured[1].params.as_ref().expect("params")["payload"],
-        json!({ "transient_id": "transient-sdk-ingest" })
     );
     server.join().expect("server joined");
 }
